Move test utilities interfaces and functions to its own file
